+++ src/OpenTelemetry.OpAmp.Client/Transport/IOpAmpTransport.cs
@@ -1,0 +1,12 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+using Google.Protobuf;
+
+namespace OpenTelemetry.OpAmp.Client.Transport;
+
+internal interface IOpAmpTransport
+{
+    Task SendAsync<T>(T message, CancellationToken token)
+        where T : IMessage<T>;
+}