+++ src/OpenTelemetry.OpAmp.Client/CHANGELOG.md
@@ -2,6 +2,5 @@
 
 ## Unreleased
 
-* Initial release of
-  `OpenTelemetry.OpAmp.Client`
-  project.
+* Initial release of `OpenTelemetry.OpAmp.Client` project.
+* Added support for OpAMP Plain Http transport.